pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "markdown>=3.10.2",
]
